requirements_bot.txt
pydantic==1.9.0
python-dotenv==0.20.0
python-telegram-bot==13.11
SQLAlchemy==1.4.32
psycopg2-binary==2.9.1